+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Implemented `SLL_ABI_AUDIT_CALL`, `SLL_ABI_AUDIT_DEINIT`, `SLL_ABI_AUDIT_INIT`, `SLL_ABI_DEINIT`, `SLL_ABI_INIT` and `SLL_ABI_NAME`
 - Implemented `sll_expand_environment_variable`
 - Implemented `sll_search_path_t`, `SLL_SEARCH_PATH_SPLIT_CHAR`, `SLL_SEARCH_PATH_FLAG_BEFORE`, `SLL_SEARCH_PATH_FLAG_AFTER`, `sll_env_path`, `sll_free_search_path`, `sll_search_path_create`, `sll_search_path_find` and `sll_search_path_from_environment`
+- Implemented `sll_weakref_clone`
 - Implemented `statistics$correlation` ([#260])
 - More examples ([#8])
 - Object creation code to build an array of string pointers: `x`